Extract progress from deck.rs
